Split MachineRequestInterface
