+++ src/Message/AbstractMachineRequest.php
@@ -6,8 +6,6 @@ namespace App\Message;
 
 abstract class AbstractMachineRequest implements MachineRequestInterface
 {
-    private int $retryCount = 0;
-
     public function __construct(
         private string $machineId,
     ) {
@@ -16,18 +14,5 @@ abstract class AbstractMachineRequest implements MachineRequestInterface
     public function getMachineId(): string
     {
         return $this->machineId;
-    }
-
-    public function getRetryCount(): int
-    {
-        return $this->retryCount;
-    }
-
-    public function incrementRetryCount(): MachineRequestInterface
-    {
-        $new = clone $this;
-        $new->retryCount++;
-
-        return $new;
     }
 }
