+++ src/Message/CreateMachine.php
@@ -6,8 +6,10 @@ namespace App\Message;
 
 use App\Model\RemoteRequestActionInterface;
 
-class CreateMachine extends AbstractMachineRequest
+class CreateMachine extends AbstractMachineRequest implements RemoteMachineRequestInterface
 {
+    use RetryableRequestTrait;
+
     public function getType(): string
     {
         return RemoteRequestActionInterface::ACTION_CREATE;
