+++ src/Message/RetryableRequestInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Message;
+
+interface RetryableRequestInterface
+{
+    public function getRetryCount(): int;
+    public function incrementRetryCount(): RetryableRequestInterface;
+}
