+++ tests/Functional/MessageDispatcher/MachineRequestMessageDispatcherTest.php
@@ -6,7 +6,7 @@ namespace App\Tests\Functional\MessageDispatcher;
 
 use App\Message\CreateMachine;
 use App\Message\MachineExists;
-use App\Message\MachineRequestInterface;
+use App\Message\RemoteMachineRequestInterface;
 use App\Message\UpdateMachine;
 use App\MessageDispatcher\MachineRequestMessageDispatcher;
 use App\Model\ProviderInterface;
@@ -49,7 +49,7 @@ class MachineRequestMessageDispatcherTest extends AbstractBaseFunctionalTest
     /**
      * @dataProvider dispatchDataProvider
      */
-    public function testDispatch(MachineRequestInterface $message, ?StampInterface $expectedDelayStamp): void
+    public function testDispatch(RemoteMachineRequestInterface $message, ?StampInterface $expectedDelayStamp): void
     {
         $this->messengerAsserter->assertQueueIsEmpty();
